Isolate `config` module with `AppConfig`
